+++ HISTORY.rst
@@ -7,6 +7,9 @@ History
 
 * Solved matplotlib DeprecationWarning.
 * 'make lint' now directly prints the log output.
+* Revised RSImage_ClusterPredictor.predict() to improve speed
+  (reduced processing time to 70-50% of the previous version).
+* Added weights attribute to prediction result.
 
 
 0.9.3 (2020-12-15)