+++ HISTORY.rst
@@ -15,6 +15,7 @@ History
 * Fixed TypeError in case the input image for the prediction has no nodata pixels and no in- and output nodata
   value is given.
 * Fixed type error in case non-kNN regressors are used for prediction.
+* Fixed Exception in case a ClassifierCollection does not have the requested key (LBA).
 
 0.9.3 (2020-12-15)
 ------------------