+++ HISTORY.rst
@@ -16,6 +16,7 @@ History
   value is given.
 * Fixed type error in case non-kNN regressors are used for prediction.
 * Fixed Exception in case a ClassifierCollection does not have the requested key (LBA).
+* Revised test_spechomo_install CI job (now uses mamba).
 
 0.9.3 (2020-12-15)
 ------------------