+++ HISTORY.rst
@@ -18,6 +18,8 @@ History
 * Fixed Exception in case a ClassifierCollection does not have the requested key (LBA).
 * Revised test_spechomo_install CI job (now uses mamba).
 * Added sphinx plugin requirements to environment_spechomo.yml.
+* Global regressors now have the correct weights and those use those from the material-specific regressors anymore.
+  The global regressor is now counted only once in the prediction.
 
 0.9.3 (2020-12-15)
 ------------------